Async Support Update - Now add AsyncChatBot to V1 - Delete AsyncV1 file
