--- src/revChatGPT/AsyncV1.py
@@ -1,444 +1,0 @@
-"""
-Standard ChatGPT
-"""
-import json
-import uuid
-from os import environ
-from os import getenv
-from os.path import exists
-import asyncio
-import httpx
-
-from typing import List, Dict
-
-import requests
-from OpenAIAuth import Authenticator, Error as AuthError
-
-BASE_URL = environ.get("CHATGPT_BASE_URL") or "https://chatgpt.duti.tech/"
-
-
-class Error(Exception):
-    """Base class for exceptions in this module."""
-
-    source: str
-    message: str
-    code: int
-
-
-class MissingFieldError(Exception):
-    def __init__(self, message, missing_fields):
-        super().__init__(message)
-        self.missing_fields = missing_fields
-
-
-class AsyncChatbot:
-    """
-    Chatbot class for ChatGPT
-    """
-
-    def __init__(
-        self,
-        config,
-        conversation_id=None,
-        parent_id=None,
-    ) -> None:
-        self.config = config
-        self.session = httpx.AsyncClient()
-        if "proxy" in config:
-            if isinstance(config["proxy"], str) is False:
-                raise Exception("Proxy must be a string!")
-            proxies = {
-                "http": config["proxy"],
-                "https": config["proxy"],
-            }
-            self.session = httpx.AsyncClient(proxies=proxies)
-        if "verbose" in config:
-            if not isinstance(config["verbose"], bool):
-                raise Exception("Verbose must be a boolean!")
-            self.verbose = config["verbose"]
-        else:
-            self.verbose = False
-        self.conversation_id = conversation_id
-        self.parent_id = parent_id
-        self.conversation_mapping: Dict = {}
-        self.conversation_id_prev_queue: List = []
-        self.parent_id_prev_queue: List = []
-        if "email" in config and "password" in config:
-            pass
-        elif "access_token" in config:
-            self.__refresh_headers(config["access_token"])
-        elif "session_token" in config:
-            pass
-        else:
-            raise Exception("No login details provided!")
-        if "access_token" not in config:
-            try:
-                self.__login()
-            except AuthError as error:
-                raise error
-
-    def __refresh_headers(self, access_token):
-        self.session.headers.clear()
-        self.session.headers.update(
-            {
-                "Accept": "text/event-stream",
-                "Authorization": f"Bearer {access_token}",
-                "Content-Type": "application/json",
-                "X-Openai-Assistant-App-Id": "",
-                "Connection": "close",
-                "Accept-Language": "en-US,en;q=0.9",
-                "Referer": "https://chat.openai.com/chat",
-            },
-        )
-
-    def __login(self):
-        if (
-            "email" not in self.config or "password" not in self.config
-        ) and "session_token" not in self.config:
-            raise Exception("No login details provided!")
-        auth = Authenticator(
-            email_address=self.config.get("email"),
-            password=self.config.get("password"),
-            proxy=self.config.get("proxy"),
-        )
-        if self.config.get("session_token"):
-            auth.session_token = self.config["session_token"]
-            auth.get_access_token()
-            if auth.access_token is None:
-                del self.config["session_token"]
-                self.__login()
-                return
-        else:
-            auth.begin()
-            self.config["session_token"] = auth.session_token
-            auth.get_access_token()
-
-        self.__refresh_headers(auth.access_token)
-
-    async def ask(
-        self,
-        prompt,
-        conversation_id=None,
-        parent_id=None,
-        timeout=360,
-        # gen_title=True,
-    ):
-        if parent_id is not None and conversation_id is None:
-            error = Error()
-            error.source = "User"
-            error.message = "conversation_id must be set once parent_id is set"
-            error.code = -1
-            raise error
-            # user-specified covid and parid, check skipped to avoid rate limit
-
-        if (
-            conversation_id is not None and conversation_id != self.conversation_id
-        ):  # Update to new conversations
-            self.parent_id = None  # Resetting parent_id
-
-        conversation_id = conversation_id or self.conversation_id
-        parent_id = parent_id or self.parent_id
-        if conversation_id is None and parent_id is None:  # new conversation
-            parent_id = str(uuid.uuid4())
-
-        if conversation_id is not None and parent_id is None:
-            if conversation_id not in self.conversation_mapping:
-                self.__map_conversations()
-            parent_id = self.conversation_mapping[conversation_id]
-        data = {
-            "action": "next",
-            "messages": [
-                {
-                    "id": str(uuid.uuid4()),
-                    "role": "user",
-                    "content": {"content_type": "text", "parts": [prompt]},
-                },
-            ],
-            "conversation_id": conversation_id,
-            "parent_message_id": parent_id,
-            "model": "text-davinci-002-render-sha"
-            if not self.config.get("paid")
-            else "text-davinci-002-render-paid",
-        }
-        # new_conv = data["conversation_id"] is None
-        self.conversation_id_prev_queue.append(
-            data["conversation_id"],
-        )  # for rollback
-        self.parent_id_prev_queue.append(data["parent_message_id"])
-
-        async with self.session.stream(
-            method="POST",
-            url=BASE_URL + "api/conversation",
-            data=json.dumps(data),
-            timeout=timeout,
-        ) as response:
-            self.__check_response(response)
-            buffer = b""
-            async for chunk in response.aiter_bytes():
-                buffer += chunk
-                while b"\n" in buffer:
-                    idx = buffer.index(b"\n")
-                    line = buffer[:idx].decode("utf-8")
-                    buffer = buffer[idx + 1 :]
-
-                    if line == "" or line is None:
-                        continue
-                    if "data: " in line:
-                        line = line[6:]
-                    if "[DONE]" in line:
-                        break
-
-                    # Replace accidentally escaped double quotes
-                    line = (
-                        line.replace('\\"', '"')
-                        .replace("\\'", "'")
-                        .replace("\\\\", "\\")
-                    )
-                    # Try parse JSON
-                    try:
-                        line = json.loads(line)
-                    except json.decoder.JSONDecodeError:
-                        continue
-                    if not self.__check_fields(line):
-                        raise Exception("Field missing. Details: " + str(line))
-
-                    message = line["message"]["content"]["parts"][0]
-                    conversation_id = line["conversation_id"]
-                    parent_id = line["message"]["id"]
-                    yield {
-                        "message": message,
-                        "conversation_id": conversation_id,
-                        "parent_id": parent_id,
-                    }
-            self.conversation_mapping[conversation_id] = parent_id
-            if parent_id is not None:
-                self.parent_id = parent_id
-            if conversation_id is not None:
-                self.conversation_id = conversation_id
-
-    def __check_response(self, response):
-        response.raise_for_status()
-
-    def __check_fields(self, data: dict) -> bool:
-        try:
-            data["message"]["content"]
-        except TypeError:
-            return False
-        except KeyError:
-            return False
-        return True
-
-    async def get_conversations(self, offset=0, limit=20):
-        """
-        Get conversations
-        :param offset: Integer
-        :param limit: Integer
-        """
-        url = BASE_URL + f"api/conversations?offset={offset}&limit={limit}"
-        response = await self.session.get(url)
-        await self.__check_response(response)
-        data = json.loads(await response.text())
-        return data["items"]
-
-    async def get_msg_history(self, convo_id, encoding="utf-8"):
-        """
-        Get message history
-        :param id: UUID of conversation
-        """
-        url = BASE_URL + f"api/conversation/{convo_id}"
-        response = await self.session.get(url)
-        if encoding is not None:
-            response.encoding = encoding
-            await self.__check_response(response)
-            data = json.loads(await response.text())
-            return data
-
-    async def gen_title(self, convo_id, message_id):
-        """
-        Generate title for conversation
-        """
-        url = BASE_URL + f"api/conversation/gen_title/{convo_id}"
-        response = await self.session.post(
-            url,
-            data=json.dumps(
-                {"message_id": message_id, "model": "text-davinci-002-render"},
-            ),
-        )
-        await self.__check_response(response)
-
-    async def change_title(self, convo_id, title):
-        """
-        Change title of conversation
-        :param convo_id: UUID of conversation
-        :param title: String
-        """
-        url = BASE_URL + f"api/conversation/{convo_id}"
-        response = await self.session.patch(url, data=f'{{"title": "{title}"}}')
-        self.__check_response(response)
-
-    async def delete_conversation(self, convo_id):
-        """
-        Delete conversation
-        :param convo_id: UUID of conversation
-        """
-        url = BASE_URL + f"api/conversation/{convo_id}"
-        response = await self.session.patch(url, data='{"is_visible": false}')
-        self.__check_response(response)
-
-    async def clear_conversations(self):
-        """
-        Delete all conversations
-        """
-        url = BASE_URL + "api/conversations"
-        response = await self.session.patch(url, data='{"is_visible": false}')
-        self.__check_response(response)
-
-    async def __map_conversations(self):
-        conversations = await self.get_conversations()
-        histories = [await self.get_msg_history(x["id"]) for x in conversations]
-        for x, y in zip(conversations, histories):
-            self.conversation_mapping[x["id"]] = y["current_node"]
-
-    def reset_chat(self) -> None:
-        """
-        Reset the conversation ID and parent ID.
-
-        :return: None
-        """
-        self.conversation_id = None
-        self.parent_id = str(uuid.uuid4())
-
-    def rollback_conversation(self, num=1) -> None:
-        """
-        Rollback the conversation.
-        :param num: The number of messages to rollback
-        :return: None
-        """
-        for _ in range(num):
-            self.conversation_id = self.conversation_id_prev_queue.pop()
-            self.parent_id = self.parent_id_prev_queue.pop()
-
-
-def get_input(prompt):
-    """
-    Multiline input function.
-    """
-    # Display the prompt
-    print(prompt, end="")
-
-    # Initialize an empty list to store the input lines
-    lines = []
-
-    # Read lines of input until the user enters an empty line
-    while True:
-        line = input()
-        if line == "":
-            break
-        lines.append(line)
-
-    # Join the lines, separated by newlines, and store the result
-    user_input = "\n".join(lines)
-
-    # Return the input
-    return user_input
-
-
-def configure():
-    """
-    Looks for a config file in the following locations:
-    """
-    config_files = ["config.json"]
-    xdg_config_home = getenv("XDG_CONFIG_HOME")
-    if xdg_config_home:
-        config_files.append(f"{xdg_config_home}/revChatGPT/config.json")
-    user_home = getenv("HOME")
-    if user_home:
-        config_files.append(f"{user_home}/.config/revChatGPT/config.json")
-
-    config_file = next((f for f in config_files if exists(f)), None)
-    if config_file:
-        with open(config_file, encoding="utf-8") as f:
-            config = json.load(f)
-    else:
-        print("No config file found.")
-        raise Exception("No config file found.")
-    return config
-
-
-async def main(config: dict):
-    """
-    Main function for the chatGPT program.
-    """
-    print("Logging in...")
-    chatbot = AsyncChatbot(
-        config,
-        conversation_id=config.get("conversation_id"),
-        parent_id=config.get("parent_id"),
-    )
-
-    def handle_commands(command: str) -> bool:
-        if command == "!help":
-            print(
-                """
-            !help - Show this message
-            !reset - Forget the current conversation
-            !config - Show the current configuration
-            !rollback x - Rollback the conversation (x being the number of messages to rollback)
-            !exit - Exit this program
-            !setconversation - Changes the conversation
-            """,
-            )
-        elif command == "!reset":
-            chatbot.reset_chat()
-            print("Chat session successfully reset.")
-        elif command == "!config":
-            print(json.dumps(chatbot.config, indent=4))
-        elif command.startswith("!rollback"):
-            # Default to 1 rollback if no number is specified
-            try:
-                rollback = int(command.split(" ")[1])
-            except IndexError:
-                rollback = 1
-            chatbot.rollback_conversation(rollback)
-            print(f"Rolled back {rollback} messages.")
-        elif command.startswith("!setconversation"):
-            try:
-                chatbot.conversation_id = chatbot.config[
-                    "conversation_id"
-                ] = command.split(" ")[1]
-                print("Conversation has been changed")
-            except IndexError:
-                print("Please include conversation UUID in command")
-        elif command == "!exit":
-            exit(0)
-        else:
-            return False
-        return True
-
-    while True:
-        prompt = get_input("\nYou:\n")
-        if prompt.startswith("!"):
-            if handle_commands(prompt):
-                continue
-
-        print("Chatbot: ")
-        prev_text = ""
-        prev_text = ""
-        async for data in chatbot.ask(prompt):
-            message = data["message"][len(prev_text) :]
-            print(message, end="", flush=True)
-            prev_text = data["message"]
-        print()
-
-
-if __name__ == "__main__":
-    print(
-        """
-        ChatGPT - A command-line interface to OpenAI's ChatGPT (https://chat.openai.com/chat)
-        Repo: github.com/acheong08/ChatGPT
-        
-        Type '!help' to show a full list of commands
-        Press enter twice to submit your question.\n
-        """,
-    )
-    asyncio.run(main(configure()))
