+++ src/revChatGPT/V1.py
@@ -1,6 +1,7 @@
 """
 Standard ChatGPT
 """
+import asyncio
 import json
 import logging
 import time
@@ -13,6 +14,7 @@ from os.path import exists
 import requests
 from OpenAIAuth import Authenticator
 from OpenAIAuth import Error as AuthError
+from httpx import AsyncClient
 
 logging.basicConfig(
     format="%(asctime)s - %(name)s - %(levelname)s - %(funcName)s - %(message)s",
@@ -86,13 +88,11 @@ class Chatbot:
 
     @logger(is_timed=True)
     def __init__(
-        self,
-        config,
-        conversation_id=None,
-        parent_id=None,
+        self, config, conversation_id=None, parent_id=None, session_client=None
     ) -> None:
         self.config = config
-        self.session = requests.Session()
+        self.session = session_client() if session_client else requests.Session()
+
         if "proxy" in config:
             if isinstance(config["proxy"], str) is False:
                 raise Exception("Proxy must be a string!")
@@ -100,21 +100,29 @@ class Chatbot:
                 "http": config["proxy"],
                 "https": config["proxy"],
             }
-            self.session.proxies.update(proxies)
+            if isinstance(self.session, AsyncClient):
+                self.session = AsyncClient(proxies=proxies)
+            else:
+                self.session.proxies.update(proxies)
         self.conversation_id = conversation_id
         self.parent_id = parent_id
         self.conversation_mapping = {}
         self.conversation_id_prev_queue = []
         self.parent_id_prev_queue = []
-        if "email" in config and "password" in config:
+
+        self.__check_credentials()
+
+    @logger(is_timed=True)
+    def __check_credentials(self):
+        if "email" in self.config and "password" in self.config:
             pass
-        elif "access_token" in config:
-            self.__refresh_headers(config["access_token"])
-        elif "session_token" in config:
+        elif "access_token" in self.config:
+            self.__refresh_headers(self.config["access_token"])
+        elif "session_token" in self.config:
             pass
         else:
             raise Exception("No login details provided!")
-        if "access_token" not in config:
+        if "access_token" not in self.config:
             try:
                 self.__login()
             except AuthError as error:
@@ -416,6 +424,203 @@ class Chatbot:
             self.parent_id = self.parent_id_prev_queue.pop()
 
 
+class AsyncChatbot(Chatbot):
+    """
+    Async Chatbot class for ChatGPT
+    """
+
+    def __init__(
+        self,
+        config,
+        conversation_id=None,
+        parent_id=None,
+    ) -> None:
+        super().__init__(
+            config=config,
+            conversation_id=conversation_id,
+            parent_id=parent_id,
+            session_client=AsyncClient,
+        )
+
+    async def ask_async(
+        self,
+        prompt,
+        conversation_id=None,
+        parent_id=None,
+        timeout=360,
+        # gen_title=True,
+    ):
+        if parent_id is not None and conversation_id is None:
+            error = Error()
+            error.source = "User"
+            error.message = "conversation_id must be set once parent_id is set"
+            error.code = -1
+            raise error
+            # user-specified covid and parid, check skipped to avoid rate limit
+
+        if (
+            conversation_id is not None and conversation_id != self.conversation_id
+        ):  # Update to new conversations
+            self.parent_id = None  # Resetting parent_id
+
+        conversation_id = conversation_id or self.conversation_id
+        parent_id = parent_id or self.parent_id
+        if conversation_id is None and parent_id is None:  # new conversation
+            parent_id = str(uuid.uuid4())
+
+        if conversation_id is not None and parent_id is None:
+            if conversation_id not in self.conversation_mapping:
+                self.__map_conversations()
+            parent_id = self.conversation_mapping[conversation_id]
+        data = {
+            "action": "next",
+            "messages": [
+                {
+                    "id": str(uuid.uuid4()),
+                    "role": "user",
+                    "content": {"content_type": "text", "parts": [prompt]},
+                },
+            ],
+            "conversation_id": conversation_id,
+            "parent_message_id": parent_id,
+            "model": "text-davinci-002-render-sha"
+            if not self.config.get("paid")
+            else "text-davinci-002-render-paid",
+        }
+        # new_conv = data["conversation_id"] is None
+        self.conversation_id_prev_queue.append(
+            data["conversation_id"],
+        )  # for rollback
+        self.parent_id_prev_queue.append(data["parent_message_id"])
+
+        async with self.session.stream(
+            method="POST",
+            url=BASE_URL + "api/conversation",
+            data=json.dumps(data),
+            timeout=timeout,
+        ) as response:
+            self.__check_response(response)
+            buffer = b""
+            async for chunk in response.aiter_bytes():
+                buffer += chunk
+                while b"\n" in buffer:
+                    idx = buffer.index(b"\n")
+                    line = buffer[:idx].decode("utf-8")
+                    buffer = buffer[idx + 1 :]
+
+                    if line == "" or line is None:
+                        continue
+                    if "data: " in line:
+                        line = line[6:]
+                    if "[DONE]" in line:
+                        break
+
+                    # Replace accidentally escaped double quotes
+                    line = (
+                        line.replace('\\"', '"')
+                        .replace("\\'", "'")
+                        .replace("\\\\", "\\")
+                    )
+                    # Try parse JSON
+                    try:
+                        line = json.loads(line)
+                    except json.decoder.JSONDecodeError:
+                        continue
+                    if not self.__check_fields(line):
+                        raise Exception("Field missing. Details: " + str(line))
+
+                    message = line["message"]["content"]["parts"][0]
+                    conversation_id = line["conversation_id"]
+                    parent_id = line["message"]["id"]
+                    yield {
+                        "message": message,
+                        "conversation_id": conversation_id,
+                        "parent_id": parent_id,
+                    }
+            self.conversation_mapping[conversation_id] = parent_id
+            if parent_id is not None:
+                self.parent_id = parent_id
+            if conversation_id is not None:
+                self.conversation_id = conversation_id
+
+    async def get_conversations(self, offset=0, limit=20):
+        """
+        Get conversations
+        :param offset: Integer
+        :param limit: Integer
+        """
+        url = BASE_URL + f"api/conversations?offset={offset}&limit={limit}"
+        response = await self.session.get(url)
+        await self.__check_response(response)
+        data = json.loads(await response.text())
+        return data["items"]
+
+    async def get_msg_history(self, convo_id, encoding="utf-8"):
+        """
+        Get message history
+        :param id: UUID of conversation
+        """
+        url = BASE_URL + f"api/conversation/{convo_id}"
+        response = await self.session.get(url)
+        if encoding is not None:
+            response.encoding = encoding
+            await self.__check_response(response)
+            data = json.loads(await response.text())
+            return data
+
+    async def gen_title(self, convo_id, message_id):
+        """
+        Generate title for conversation
+        """
+        url = BASE_URL + f"api/conversation/gen_title/{convo_id}"
+        response = await self.session.post(
+            url,
+            data=json.dumps(
+                {"message_id": message_id, "model": "text-davinci-002-render"},
+            ),
+        )
+        await self.__check_response(response)
+
+    async def change_title(self, convo_id, title):
+        """
+        Change title of conversation
+        :param convo_id: UUID of conversation
+        :param title: String
+        """
+        url = BASE_URL + f"api/conversation/{convo_id}"
+        response = await self.session.patch(url, data=f'{{"title": "{title}"}}')
+        self.__check_response(response)
+
+    async def delete_conversation(self, convo_id):
+        """
+        Delete conversation
+        :param convo_id: UUID of conversation
+        """
+        url = BASE_URL + f"api/conversation/{convo_id}"
+        response = await self.session.patch(url, data='{"is_visible": false}')
+        self.__check_response(response)
+
+    async def clear_conversations(self):
+        """
+        Delete all conversations
+        """
+        url = BASE_URL + "api/conversations"
+        response = await self.session.patch(url, data='{"is_visible": false}')
+        self.__check_response(response)
+
+    async def __map_conversations(self):
+        conversations = await self.get_conversations()
+        histories = [await self.get_msg_history(x["id"]) for x in conversations]
+        for x, y in zip(conversations, histories):
+            self.conversation_mapping[x["id"]] = y["current_node"]
+
+    def __check_fields(self, data: dict) -> bool:
+        return self._Chatbot__check_fields(data)
+
+    def __check_response(self, response):
+        response.raise_for_status()
+
+
 @logger(is_timed=False)
 def get_input(prompt):
     """
@@ -465,12 +670,12 @@ def configure():
 
 
 @logger(is_timed=False)
-def main(config: dict):
+async def main(config: dict):
     """
     Main function for the chatGPT program.
     """
     print("Logging in...")
-    chatbot = Chatbot(
+    chatbot = AsyncChatbot(
         config,
         conversation_id=config.get("conversation_id"),
         parent_id=config.get("parent_id"),
@@ -531,7 +736,7 @@ def main(config: dict):
 
         print("Chatbot: ")
         prev_text = ""
-        for data in chatbot.ask(
+        async for data in chatbot.ask_async(
             prompt,
         ):
             message = data["message"][len(prev_text) :]
@@ -550,4 +755,4 @@ if __name__ == "__main__":
     )
     print("Type '!help' to show a full list of commands")
     print("Press enter twice to submit your question.\n")
-    main(configure())
+    asyncio.run(main(configure()))
